Initial commit of git-version module
This can be used by other modules to retrieve a string describing the
version of the current git repo, based on `git describe`.

The version is expected to be included in the module's state some how,
so that it's possible to determine which version of a module has been
deployed.

This relies on the user having Python and Git installed.

INFRA-720