Override version with `TF_MODULE_VERSION`, if set
This allows us to override the version returned by setting the
`TF_MODULE_VERSION` environment variable.  This is useful in situations
where the Git repo isn't available at apply-time.

The name of the environment variable to check defaults to
`TF_MODULE_VERSION`, but can be set to a different value using the
`version_environment_variable` variable.

AWS-582