Return `version_override` variable value if set
This allows users of the module to override the version value to be
returned by passing it in as a variable.

This can be helpful if the version can be passed to the parent module
via a variable, saving the parent module from needing to implement logic
to optionally call the module.

AWS-582